refactor: postCard 테스트 수정
